Revert "Add logic for Allocation structs for max usage"
This reverts commit 039ccfe825780dd581f59b9cded3f30b5e01a805.

This logic for aggregating vs. accumulating maximum usage does not work.
Essentially, determining overall maximum of a set of weighted intervals
is a fun algorithms problem and requires us to keep around a set of
intervals and maximums for every sub-allocation that a cumulative
Allocation is composed of. That's too heavy for the Allocation type, so
we will instead just keep max at the raw level (next commit) and zero
it out as soon as any operation like .add() is called on Allocation.