Emit kube_pod_labels with all label names
In KSM v2, kube_pod_labels was restricted to only emit labels
if they are explicitly whitelisted. See
https://github.com/kubernetes/kube-state-metrics/issues/1270#issuecomment-712986441

We need labels so we can match pods with services, among other things.

This commit enables cost-model to emit kube_pod_labels as if it were
pre-KSM v2. This should not cause problems if KSM v1 is deployed because
we are just emitting the same data (which will have a different job label)
and it isn't used for things like sum/rate/etc. queries.