Add Start, End, and Minutes to AllocationSetRange
These methods are useful because they enable more accurate rate
calculations. Instead of relying on the ASR's Window (which generally
maps to the _query_ window) these are computed from the Allocations
in the ASR. So if Kubecost has been running for only a day and 30d
of data is queried, the AllocationSetRange's Window will be ~30d but
the new Start(), End(), and Minutes() will reflect the data within
the ASR.

At some point we may want to modify the API response to include metadata
like this for AllocationSetRanges and AllocationSets.

Tested with the unit tests!