Alloc.add handles changes in only controller name
The new AddDifferentController test failed before the change to
Allocation.add and outlines the situation. We want to be able
to handle a situation that the Kubernetes spec does not expect
but occurs in the wild: a pod whose controller's name changes
but nothing else does. We'd like the two _technically_ separate
allocations that this generates to be added as if they were the
same without empty-stringing the controller name field. We opted
to pick the alphabetically first string but another strategy
could be used instead.