Avoid nil panic on non-GKE GCP clusters
A user running K3s on GCP is encountering a nil panic
caused by our parsing logic for the node.kubernetes.io/instance-type
label. On GCP, we expect GCP node type formats (xx-xx-xx) with
two dashes. The K3s cluster sets this value as just "k3s" in the
user's cluster.

By adding a guard on the length of the parsed result we avoid the panic.