Emit the node_gpu_hourly_cost metric
The GaugeVec for recording node_gpu_hourly_cost is correctly initialized
but is never written to. This behavior can be observed by deploying
Kubecost and visiting /model/metrics. You will note that the
node_gpu_count metric exists but the node_gpu_hourly_cost metric does not.

This is causing GPU cost to be 0 throughout the Kubecost application,
even if GPUs are present on a Node and their count is correctly
recorded (see the Assets API).

Fixed by emitting GPU cost!