Regenerate codecs with backwards-compat support
Bingen supports backwards compatibility (1). This commit doesn't add
any fields, so it should be safe to keep the schema version as v15.
All data should effectively read and write the same, just with support
for reading v<=15 schemas. Migration of old schema files to the new
file location will be part of a different PR.

1. https://github.com/kubecost/bingen/pull/7