Implement AllocationFilter.Equals()
Introduce support for logical equivalence between values of the
AllocationFilter type. This equivalence is for checking e.g. that two
ORs with filters in different orders are, at their core, filtering
the same thing: (or a b c) = (or b c a)

This is motivated by the need to check filter equivalence in unit tests
that don't have stable filter construction, like when constructing a
filter while iterating over a map.

Signed-off-by: Michael Dresser <michaelmdresser@gmail.com>