Make usage of Disk resilient to absence of metrics
Before this change, we were unable to distinguish between these
scenarios:
A. Usage metric exists in Prometheus and has a value of 0
B. Usage metric does not exist in Prometheus

Because we were using a float64, situation B would cause usage to be
unset, making it the Go default value: 0. This is bad data because the
usage is probably not 0.

With this commit, the Go default value becomes nil, making it easy to
distinguish between situations A and B.

Signed-off-by: Michael Dresser <michaelmdresser@gmail.com>