Remove HPA data from cluster cache
Data is not used anywhere by OpenCost or Kubecost. It is causing a
problem when installing on K8s v1.25 because we are using v2beta1 which
was previously deprecated and is now removed in v1.25.

While we could bump this to v2, that would remove support for K8s
versions before v1.23 (the release in which autoscaling/v2 went GA).

Opting to remove this data entirely for the time being. We can choose
which version of the autoscaling API to support here sometime in the
future, once we know that we need the data.

Signed-off-by: Michael Dresser <michaelmdresser@gmail.com>