fix(query): Use seconds with isRequestStepAligned and alignWindow
Caught two bugs due to mismatch in seconds and milliseconds
when calculating the updated values for aligning the window.

`isRequestStepAligned` will now convert the step to seconds and then check
if the modulo of the start and end time(represented as unix time stamps)
is zero. If either of the values is not aligned, return false. Otherwise
will return true.

`alignWindow` will be called if the result of `isRequestStepAligned` is
false. This method accepts the start, end and step times. It will
convert each number to seconds and then truncate the amount time
proportional to the step function. The current implementation will round
always round down. IE, if the step is an hour then the minutes and
seconds will be removed. See the unit tests for some examples of how the
times will be aligned.

Adds basic unit tests for both methods.

Signed-off-by: pokom <mark.poko@grafana.com>