feat(config): Expose DisableCache as an environment variable
This adds the ability to set the value for `disableCache` via an
environment variable `DISABLE_CACHE`. The default value is set to false.
This will influence whether ComputeAggregateCostModel will use
data if it's found in the cache.

Add unit tests to ensure the default value for `IsCacheDisabled` is
false and that it can be overidden.

- Relates to #1472

Signed-off-by: pokom <mark.poko@grafana.com>