feat(build): Add version and revision as labels
This adds a module `pkg/build` that is responsible for setting
information about the build and providing a method `GetBuildInfo` that
will return an object with associated build information.

The way that it works is at build time, we need to provide a set of
[extldflags](https://pkg.go.dev/cmd/link) which will set the variables defined
in `pkg/build/build.go`. There is an associated Makefile that will
populate the following variables:
- IMAGE_TAG => Version
- GIT_REVISION => Revision
- GIT_BRANCH => Branch
- GIT_LAST_COMMIT_DATE => BuildDate

These are then exported via the `opencost_build_info` metric as labels.

Signed-off-by: pokom <mark.poko@grafana.com>