Switch CPU max query to existing irate rec rule
The query before this commit is broken; it looks like a max but is
actually an average (or very close to) due to the usage of "rate" before
maxing. The rate reduces the resolution to the first and last datapoint.

The recording rule handles this for us. At the time of this commit, the
recording rule is set in the Helm values:

- expr: avg(irate(container_cpu_usage_seconds_total{container_name!="POD", container_name!=""}[5m])) by (container_name,pod_name,namespace)
  record: kubecost_container_cpu_usage_irate

Signed-off-by: Michael Dresser <michaelmdresser@gmail.com>