Move pricesheetDownloader to azurepricesheet.Downloader
This avoids polluting the cloud package.

Pass a function to convert meter info into pricings in as config for
the downloader so we can share the convertMeterToPricings function
between the rate card and price sheet paths.

Similarly, since AzurePricing is defined in the cloud package, make
Downloader generic over that type. I'm not sure about this, it seems a
little clever - is there a sensible place to move AzurePricing to so
it can be imported in both pkg/cloud and pkg/cloud/azurepricesheet?

Signed-off-by: Christian Muirhead <christian.muirhead@microsoft.com>