refactor(cloud): Move aws and gcp to dedicated modules
Following the move by Azure folks to refactor the azure module in #1869,
I am refactoring AWS and GCP into their own dedicated folders. This was
for the most part straight forward. The one caveat is that the
customprovider had hard dependencies on AWS resources. This didn't feel
right to me, so instead of the two sharing certain structs, I opted to
copy over the pvKey implementation to customprovider and a regex.

Signed-off-by: pokom <mark.poko@grafana.com>