Implement Equal() for Overhead and test equality
The full test is of (*Node).Equal() which necessitated a change to
equality for Breakdown and Properties.

It is odd to say that two things which are nil pointers of the same type
are unequal (the old behavior). I think this is an acceptable change.
The risk revolves around any functionality which expects Equal() to fail
on two nil pointers -- I sincerely hope that nothing behaves this way.

Signed-off-by: Michael Dresser <michaelmdresser@gmail.com>