Enable (*FileStorage).ListDirectories to list symlinks
Done by making DirFileToStargeInfo follow symlinks to determine if the
target is a directory, then returning the _unfollowed_ path as a valid
directory so the path will be correctly constructed relative to the
symlink, rather than relative to the target. The purpose is only to
ensure that the target is in fact a directory.

Signed-off-by: Michael Dresser <michaelmdresser@gmail.com>